feat(jpa): change findAllBetweenId query method parameters type

+++ springboot-jpa/src/main/java/com/crhistianm/springboot/jpa/springboot_jpa/repositories/PersonRepository.java
@@ -25,7 +25,7 @@ public interface PersonRepository extends CrudRepository<Person, Long>{
     List<Person> findAllBetweenName(String c1, String c2);
 
     @Query("select p from Person p where p.id between ?1 and ?2")
-    List<Person> findAllBetweenId(Integer id1, Integer id2);
+    List<Person> findAllBetweenId(Long id1, Long id2);
 
     @Query("select p.id, upper(p.name), lower(p.lastname), upper(p.programmingLanguage) from Person p")
     List<Object[]> findPersonDataListCase();
